requirements.txt
flask
yt-dlp
discord.py
PyNaCl
